medical_consultation_ml/requirements.txt
# Medical Consultation Transcription & Summarization Project Requirements

# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
openai-whisper>=20231117
faster-whisper>=0.10.0

# NLP & Text Processing
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0
sentencepiece>=0.1.99

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# Data Handling
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Web Interface
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# API & Integration
requests>=2.31.0
aiohttp>=3.9.0

# Storage & Database
sqlalchemy>=2.0.0

# Logging & Monitoring
loguru>=0.7.2

# Optional: GPU acceleration
# accelerate>=0.24.0
# bitsandbytes>=0.41.0

# Development Tools (Optional)
# pytest>=7.4.0
# black>=23.11.0
# flake8>=6.1.0
